backend/pyproject.toml
[project]
name = "todoevolve-backend"
version = "0.1.0"
description = "TodoEvolve Backend API - FastAPI with SQLModel"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.12",
    "google-generativeai>=0.8.6",
    "mcp>=1.26.0",
    "langgraph>=1.0.7",
    "openai>=2.16.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "dapr-client>=1.13.0",
    "aiokafka>=0.12.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
